requirements.txt
flask==2.0.1
pillow==9.5.0
torch==1.12.1
torchvision==0.13.1
transformers==4.30.2
pdfplumber==0.9.0
numpy
pandas
kaggle
requests
